修改静态文件
ios增加版本和版本号

+++ src/main/java/com/lqkj/web/cmccr2/modules/menu/service/MenuService.java
@@ -50,7 +50,7 @@ public class MenuService {
         for (Long i : id) {
             CcrMenu menu = menuDao.getOne(i);
 
-            if (!menu.getType().equals(CcrMenu.IpsMenuType.embed)) {
+            if (menu.getType()!= null && !menu.getType().equals(CcrMenu.IpsMenuType.embed)) {
                 menuDao.delete(menu);
             }
         }
